Change scraper class; reflect the Austin utility
The original class name from the parent repo was derived from the fact
that the Code for Kentuckiana folks were scraping from the LG&E/KU
outage map, so the class name reflected this fact.  It would be a bit
weird to keep the original `lgeku` naming even when changing the class
configuration to scrape Austin Energy.