Fix for cluster layer name
Yesterday I mentioned that while trying to reproduce the Code for
Kentuckiana setup without Austin-specific adaptations that I was unable
to get the scraper to work even then.  The URLs for the relevant data
wrt to Austin-specific details look like e.g.

/cluster-data/022/20d293c6-08fd-41b3-b96b-d2f522c74990/25273d62-9820-44a2-8fcb-f95dcef84a6c/public/cluster-1/0231301220.json

Because the scraper was using the "layerName" field, it was constructing
incorrect URLs, resulting in 404s, and it turns out that Code for
Kentuckiana had been using a locally modified version with this change:

<https://github.com/codeforkyana/power-outage-data/commit/bc8c5647fc79ad18f6c5078726e895d20286c18f>

With this fix, I was able replicate both their original setup, as well
as successfully scrape after changing the scraper to match the Austin
Energy configuration.