Add mostly finished scraper implementation
Two outstanding items in order of priority (most to least):
1. actual polyline decoding
2. get rid of the need for the bounding box tiling hack

The former can be yanked from Google's freely licensede version.

The latter is not strictly necessary to get us actually operational--
with the test data, at least--but should be addressed eventually (in the
event that Austin Energy's service area expands).