Add real ability to specify khar2json output file
Only makes sense for the CLI.  So no special support needed in the
browser support module.

(Although, now that I think about it, there may be some people who have
everything set to automatically save to their downloads directory and
so will never be prompted by their browser for the file name and
directory to save to...  This is actually most browsers' default
configuration nowadays, I think.  In that case, they should seriously
consider changing their browser settings.  That's a terrible default to
begin with, for other reasons entirely.)